backend/requirements.txt
# Web framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# Database
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
alembic>=1.13.0

# Scheduling
apscheduler>=3.10.0

# LLM
anthropic>=0.25.0
openai>=1.0.0

# Gmail
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.120.0

# Notion
notion-client>=2.2.0

# Market data
vnstock>=0.2.9
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Charts
matplotlib>=3.8.0

# Utils
python-dotenv>=1.0.0
pydantic-settings>=2.0.0
python-dateutil>=2.8.0
Pillow>=10.0.0
httpx>=0.27.0
pyyaml>=6.0
lunardate>=0.2.0

# Dev only
pytest>=8.0.0
pytest-asyncio>=0.23.0
